Renames: NumberTheory -> Numeric.
