+++ Numeric/Factors.hpp
@@ -12,27 +12,12 @@
 #include <numeric>
 #include <utility>
 
-namespace NumberTheory {
+namespace Numeric {
 	/*!
 		@returns The greatest common divisor of @c a and @c b.
 	*/
 	template<class T> T greatest_common_divisor(T a, T b) {
-#if __cplusplus > 201402L
 		return std::gcd(a, b);
-#else
-		if(a < b) {
-			std::swap(a, b);
-		}
-
-		while(1) {
-			if(!a) return b;
-			if(!b) return a;
-
-			T remainder = a%b;
-			a = b;
-			b = remainder;
-		}
-#endif
 	}
 
 	/*!
@@ -40,10 +25,7 @@ namespace NumberTheory {
 		common divisor.
 	*/
 	template<class T> T least_common_multiple(T a, T b) {
-		if(a == b) return a;
-
-		T gcd = greatest_common_divisor<T>(a, b);
-		return (a / gcd) * (b / gcd) * gcd;
+		return std::lcm(a, b);
 	}
 }
