+++ Storage/Disk/Track/PCMTrack.cpp
@@ -7,7 +7,7 @@
 //
 
 #include "PCMTrack.hpp"
-#include "../../../NumberTheory/Factors.hpp"
+#include "../../../Numeric/Factors.hpp"
 #include "../../../Outputs/Log.hpp"
 
 using namespace Storage::Disk;
